split datafusion-object-store module
