+++ datafusion/Cargo.toml
@@ -63,6 +63,7 @@ datafusion-common = { path = "../datafusion-common", version = "7.0.0", features
 datafusion-expr = { path = "../datafusion-expr", version = "7.0.0" }
 datafusion-jit = { path = "../datafusion-jit", version = "7.0.0", optional = true }
 datafusion-physical-expr = { path = "../datafusion-physical-expr", version = "7.0.0" }
+datafusion-storage = { path = "../datafusion-storage", version = "7.0.0" }
 futures = "0.3"
 hashbrown = { version = "0.12", features = ["raw"] }
 lazy_static = { version = "^1.4.0" }
