+++ datafusion/src/catalog/schema.rs
@@ -24,9 +24,10 @@ use std::collections::HashMap;
 use std::sync::Arc;
 
 use crate::datasource::listing::{ListingTable, ListingTableConfig};
-use crate::datasource::object_store::{ObjectStore, ObjectStoreRegistry};
+use crate::datasource::object_store_registry::ObjectStoreRegistry;
 use crate::datasource::TableProvider;
 use crate::error::{DataFusionError, Result};
+use datafusion_storage::object_store::ObjectStore;
 
 /// Represents a schema, comprising a number of named tables.
 pub trait SchemaProvider: Sync + Send {
@@ -250,8 +251,8 @@ mod tests {
     use crate::catalog::schema::{
         MemorySchemaProvider, ObjectStoreSchemaProvider, SchemaProvider,
     };
+    use crate::datafusion_storage::object_store::local::LocalFileSystem;
     use crate::datasource::empty::EmptyTable;
-    use crate::datasource::object_store::local::LocalFileSystem;
     use crate::execution::context::SessionContext;
 
     use futures::StreamExt;
