+++ datafusion/src/datasource/file_format/json.rs
@@ -30,12 +30,12 @@ use futures::StreamExt;
 
 use super::FileFormat;
 use super::FileScanConfig;
-use crate::datasource::object_store::{ObjectReader, ObjectReaderStream};
 use crate::error::Result;
 use crate::logical_plan::Expr;
 use crate::physical_plan::file_format::NdJsonExec;
 use crate::physical_plan::ExecutionPlan;
 use crate::physical_plan::Statistics;
+use datafusion_storage::object_store::{ObjectReader, ObjectReaderStream};
 
 /// The default file extension of json files
 pub const DEFAULT_JSON_EXTENSION: &str = ".json";
@@ -102,13 +102,11 @@ mod tests {
     use super::*;
     use crate::prelude::{SessionConfig, SessionContext};
     use crate::{
-        datasource::{
-            file_format::FileScanConfig,
-            object_store::local::{
-                local_object_reader, local_object_reader_stream,
-                local_unpartitioned_file, LocalFileSystem,
-            },
+        datafusion_storage::object_store::local::{
+            local_object_reader, local_object_reader_stream, local_unpartitioned_file,
+            LocalFileSystem,
         },
+        datasource::file_format::FileScanConfig,
         physical_plan::collect,
     };
