+++ datafusion/src/physical_plan/limit.rs
@@ -428,7 +428,7 @@ mod tests {
     use common::collect;
 
     use super::*;
-    use crate::datasource::object_store::local::LocalFileSystem;
+    use crate::datafusion_storage::object_store::local::LocalFileSystem;
     use crate::physical_plan::coalesce_partitions::CoalescePartitionsExec;
     use crate::physical_plan::common;
     use crate::physical_plan::file_format::{CsvExec, FileScanConfig};
