+++ datafusion/src/physical_plan/sorts/sort.rs
@@ -595,7 +595,7 @@ async fn do_sort(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::datasource::object_store::local::LocalFileSystem;
+    use crate::datafusion_storage::object_store::local::LocalFileSystem;
     use crate::execution::context::SessionConfig;
     use crate::execution::runtime_env::RuntimeConfig;
     use crate::physical_plan::coalesce_partitions::CoalescePartitionsExec;
